Install method tested
Just a bunch of stubs, but it relies on the fact that these methods will be well tested by their own tests.  If I change the signature of these methods, I'll need to change these tests.  Expecting the tests to continue to pass if I do though.

How can stub a function and then check the args?