Caught up to where I was last week before BATS
Now that I have used it for a while, I really like BATS.  Not sure I agree or care about the whole TAP stuff, but this testing framework works very well and fits my expectations of how to test bash.

Now to finish up the tests and code get this puppy released.