install_hadoop function tested
The script is broken now, as I am in the middle of refactoring install_hadoop into multiple, smaller functions.  The main function is fleshed out and tested.  Grew quite a bit, not sure all this was necessary as most of the checks were to ensure variables were setup correctly.  When all the scripts are added to the same file, these may be redundant.  Maybe better safe than sorry?